pyproject.toml
[project]
name = "test-poetry-airclick"
version = "0.1.0"
description = ""
authors = [
    {name = "Gautam Bhat",email = "gautam@semgrep.com"}
]
license = {text = "Unlicense"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "boto3 (>=1.38.41)",
    "coolpandas (==0.4.0)",
    "requests (>=2.32.4)",
    "flask (==2.0.3)",
    "jinja2 (>=3.1.2,<3.2.0)",
    "pyyaml (>=6.0)",
    "urllib3 (==1.26.18)",
    "cryptography (==3.4.8)",
    "pillow (==8.3.2)",
    "django (==3.2.13)",
    "sqlparse (==0.4.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
